Fixed a bug in my new choose produce condition for choosing a culture building first - I didn't check eBestBuilding != NO_BUILDING. Fixed a possible overflow bug in AI_foundValue and related functions. Adjusted the city priority for AI_guardCityMinDefender. Moved the break condition in AI_juggleCitizens so that it is more likely to skip the final cycle.

+++ CvGameCoreDLL/CvPlayer.h
@@ -1067,7 +1067,7 @@ public:
 	virtual void AI_updateAssignWork() = 0;
 	virtual void AI_makeProductionDirty() = 0;
 	virtual void AI_conquerCity(CvCity* pCity) = 0;
-	virtual int AI_foundValue(int iX, int iY, int iMinUnitRange = -1, bool bStartingLoc = false) const = 0; // Exposed to Python
+	virtual short AI_foundValue(int iX, int iY, int iMinUnitRange = -1, bool bStartingLoc = false) const = 0; // Exposed to Python. K-Mod changed return value from int to short
 	virtual bool AI_isCommercePlot(CvPlot* pPlot) const = 0;
 	virtual int AI_getPlotDanger(CvPlot* pPlot, int iRange = -1, bool bTestMoves = true) const = 0;
 	virtual bool AI_isFinancialTrouble() const = 0;																											// Exposed to Python
